// types/project.ts
export type Project = {
  id: string;
  projectType: string; // Proje Türü
  projectName: string; // Proje Adı / Project Name
  projectYear: string; // Proje Yılı / Project Year
  region: string; // Bölge
  city: string; // Şehir
  country: string; // Ülke
  totalArea: string; // Toplam Alan / Total Area
  client: string; // İşveren / Client
  services: string; // Hizmetlerimiz / Services
  description: string; // Açıklama
};